LayoutTests/http/tests/ssl/media-stream/get-user-media-nested-expected.txt

CONSOLE MESSAGE: line 52: The top-level frame has prevented a document with a different security origin from calling getUserMedia.
Tests that getUserMedia fails when the top level document and iframe do not have the same domain.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


URL: https://127.0.0.1:8443/ssl/media-stream/get-user-media-nested.html


--------
Frame: '<!--frame1-->'
--------
URL: https://localhost:8443/ssl/media-stream/resources/get-user-media-frame.html?fail;3
 

--------
Frame: '<!--frame2-->'
--------
URL: https://localhost:8443/ssl/media-stream/resources/get-user-media-frame.html?fail;2
 

--------
Frame: '<!--frame3-->'
--------
URL: https://localhost:8443/ssl/media-stream/resources/get-user-media-frame.html?fail;1

PASS getUserMedia failed with error: "NotAllowedError: The request is not allowed by the user agent or the platform in the current context, possibly because the user denied permission."

TEST COMPLETE
